backend/requirements.txt
Django>=4.0
djangorestframework>=3.14
django-cors-headers>=3.13
geopy>=2.3
reportlab>=3.6
polyline>=2.0
requests>=2.28
gunicorn>=21.2.0
